+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to the ngx Monorepo Library Builder project will be document
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2020-08-05
+### Added 
+- Extra info regarding childprocess errors
+- More specific logs
+
+## [0.3.1] - 2020-08-04
+### Fixed 
+- Handling of childprocess errors
+
+## [0.3.0] - 2020-06-22
+### Added 
+- New 'single' method to public  api of Builder
+
 ## [0.2.1] - 2020-06-22
 ### Fixed
 - Issue with missing CHANGELOG.md file